Improved packet validation
instead of checking on each received packet if it matches to ALL of our
widgets, we just check if the basic {P0|whatever here} structure is
present. And once we extrapolate the ID, and detect the corresponding
widget we only once check if packet is valid for that particular widget.